LayoutTests/imported/w3c/web-platform-tests/webrtc/RTCPeerConnection-iceConnectionState-disconnected.https-expected.txt


PASS ICE goes to disconnected if the other side goes away
FAIL ICE restart when ICE is disconnected results in checking, then connected assert_equals: expected "checking" but got "connected"
